feat: finalize backend updates for library system

+++ backend/services/bookService.js
@@ -1,3 +1,4 @@
+const mongoose = require('mongoose');
 const Book = require('../models/book');
 
 const getAllBooks = async () => {
@@ -5,6 +6,7 @@ const getAllBooks = async () => {
 };
 
 const getBookById = async (id) => {
+  if (!mongoose.Types.ObjectId.isValid(id)) return null;
   return await Book.findById(id);
 };
 
@@ -37,8 +39,36 @@ const createBook = async (
 };
 
 const deleteBook = async (id) => {
+  if (!mongoose.Types.ObjectId.isValid(id)) return null;
   return await Book.findByIdAndDelete(id);
 };
 
 const updateBook = async (id, updatedData) => {
-  const updatedBooks = await Book.findByIdAndUpdate(id, update
+  if (!mongoose.Types.ObjectId.isValid(id)) return null;
+
+  try {
+    const updatedBook = await Book.findByIdAndUpdate(id, updatedData, {
+      new: true,
+      runValidators: true,
+      context: 'query'
+    });
+    return updatedBook;
+  } catch (err) {
+    if (err && err.code === 11000) {
+      const field = Object.keys(err.keyValue || {})[0] || 'field';
+      const message = `Duplicate value for ${field}. Please provide a unique value.`;
+      const e = new Error(message);
+      e.status = 409;
+      throw e;
+    }
+    throw err;
+  }
+};
+
+module.exports = {
+  getAllBooks,
+  getBookById,
+  createBook,
+  deleteBook,
+  updateBook
+};
